update tworoom/keyroom file structure
