+++ stable_worldmodel/envs/two_room/__init__.py
@@ -1,4 +1,10 @@
+from .doorkey_env import TwoRoomDoorKeyEnv
 from .env import TwoRoomEnv
-from .expert_policy import ExpertPolicy
+from .expert_policy import DoorKeyExpertPolicy, ExpertPolicy
 
-__all__ = ['ExpertPolicy', 'TwoRoomEnv']
+__all__ = [
+    'DoorKeyExpertPolicy',
+    'ExpertPolicy',
+    'TwoRoomDoorKeyEnv',
+    'TwoRoomEnv',
+]
