+++ stable_worldmodel/envs/two_room/doorkey_env.py
