--- stable_worldmodel/envs/two_room_doorkey/__init__.py
@@ -1,4 +1,0 @@
-from .env import TwoRoomDoorKeyEnv
-from .expert_policy import DoorKeyExpertPolicy
-
-__all__ = ['DoorKeyExpertPolicy', 'TwoRoomDoorKeyEnv']
